ai-proctoring/requirements-free.txt
# 🆓 FREE AI PROCTORING REQUIREMENTS
# No OpenAI API required - uses local OpenCV models

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision (Free & Open Source)
opencv-python==4.8.1.78
numpy==1.24.3

# Image Processing
Pillow==10.0.1

# Optional: Enhanced ML Models (Free)
# Uncomment if you want more advanced features
# scikit-learn==1.3.2
# tensorflow==2.15.0  # For neural network models
# torch==2.1.0  # For PyTorch models
# transformers==4.35.2  # For Hugging Face models (free local models)

# Utilities
python-dotenv==1.0.0
logging==0.4.9.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
